fix: update import paths after component reorganization
- Update SidebarToggle import paths in base components
- Update StorySidebar import paths after folder reorganization
- Update VocabularySection import paths after moving to story folder
- Ensure all components work correctly with new folder structure
